bug: fix sessions dates not displaying
