+++ front/public/items.json
@@ -1,12 +1,12 @@
 {
     "items": [
-        { "id": 1, "data": "20/04/2024", "period":"0-5", "df": 0, "dc": 0, 
+        { "id": 1, "date": "20/04/2024", "period":"0-5", "df": 0, "dc": 0, 
         "text":"O paciente esteve tranquilo"},
-        { "id": 2, "data": "21/04/2024", "period":"5-10", "df": 6, "dc": 5,
+        { "id": 2, "date": "21/04/2024", "period":"5-10", "df": 6, "dc": 5,
         "text":"O paciente demonstrou estresse moderado"},
-        { "id": 3, "data": "22/04/2024", "period":"10-15", "df": 10, "dc": 10, 
+        { "id": 3, "date": "22/04/2024", "period":"10-15", "df": 10, "dc": 10, 
         "text":"O paciente revelou profunda ansiedade sobre o tema"},
-        { "id": 4, "data": "29/04/2024", "period":"15-20", "df": 4, "dc": 0, 
+        { "id": 4, "date": "29/04/2024", "period":"15-20", "df": 4, "dc": 0, 
         "text":"O paciente revelou conturbações sobre o tema."}
   ]
