+++ front/vite.config.js
@@ -12,7 +12,7 @@ export default defineConfig({
     alias: {
       '@': fileURLToPath(new URL('./src', import.meta.url))
     }
-  }
+  },
   server: {
     port: 5173,
   },
